dadacat-twilio/requirements.txt
# Core dependencies
boto3>=1.26.0
twilio>=7.16.0
openai>=1.0.0
python-dotenv>=1.0.0
flask>=2.2.3
requests>=2.28.2

# AWS specific
aws-lambda-powertools>=2.13.0

# Error handling
backoff>=2.2.1

# Analytics
matplotlib>=3.7.1
pandas>=2.0.0

# Testing
pytest>=7.3.1
moto>=4.1.3
pytest-cov>=4.1.0
pytest-mock>=3.10.0

# Development
flake8>=6.0.0
black>=23.3.0
isort>=5.12.0
mypy>=1.2.0